Remove GameMode.hpp and fix GetNetMode hook

+++ SDK/Unreal/World.hpp
@@ -15,7 +15,7 @@ public:
 
     void Listen();
 
-    static uint8 GetNetModeHook(UWorld* World)
+    static int64 GetNetModeHook(UWorld* World)
     {
         return 1;
     }
